fix(logging): fix logging not keeping it's log level in other files
