+++ src/so_vits_svc_fork/gui.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import json
-from logging import getLogger
 from pathlib import Path
 
 import librosa
@@ -11,14 +10,12 @@ import torch
 from pebble import ProcessFuture, ProcessPool
 from tqdm.tk import tqdm_tk
 
-from .__main__ import init_logger
+from so_vits_svc_fork.logger import LOG
+
 from .utils import ensure_hubert_model
 
 GUI_DEFAULT_PRESETS_PATH = Path(__file__).parent / "default_gui_presets.json"
 GUI_PRESETS_PATH = Path("./user_gui_presets.json").absolute()
-init_logger()
-
-LOG = getLogger(__name__)
 
 
 def play_audio(path: Path | str):
