+++ src/so_vits_svc_fork/vdecoder/hifigan/nvSTFT.py
@@ -1,7 +1,6 @@
 import os
 
 os.environ["LRU_CACHE_CAPACITY"] = "3"
-from logging import getLogger
 
 import librosa
 import numpy as np
@@ -10,7 +9,7 @@ import torch
 import torch.utils.data
 from librosa.filters import mel as librosa_mel_fn
 
-LOG = getLogger(__name__)
+from so_vits_svc_fork.logger import LOG
 
 
 def load_wav_to_torch(full_path, target_sr=None, return_empty_on_exception=False):
